Tercera practica integradora

+++ src/services/products.service.js
@@ -27,12 +27,13 @@ class ProductsService {
         return productPayloadDTO
     }
 
-    async createProduct(productPayload, files){
+    async createProduct(productPayload, files, owner){
         const { title, description, code, stock, price, category } = productPayload
         if(!title || !description || !code || !stock || !price || !category){
             throw new HttpError('Please include all the required fields', HTTP_STATUS.BAD_REQUEST)
         }
-        const productPayloadDTO = new AddProductDTO(productPayload, files)
+
+        const productPayloadDTO = new AddProductDTO(productPayload, files, owner)
         const newProduct = productsDao.add(productPayloadDTO)
         return newProduct
     }
@@ -50,9 +51,16 @@ class ProductsService {
         return updatedProduct
     }
 
-    async deleteProduct(pid){
+    async deleteProduct(pid, user){
         if(!pid){
             throw HttpError('Please specify a product ID', HTTP_STATUS.BAD_REQUEST)
+        }
+        const product = await productsDao.getById(pid)
+        if(!product){
+            throw new HttpError('Product not found', HTTP_STATUS.NOT_FOUND)
+        }
+        if(user.role === 'premium' && user.email !== product.owner){
+            throw new HttpError("Only product's owner can delete this resource", HTTP_STATUS.FORBIDDEN)
         }
         const deletedProduct = await productsDao.delete(pid)
         return deletedProduct
